pyproject.toml
[project]
name = "joefg-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.4",
]

[tool.pytest.ini_options]
pythonpath = [
    ".", "app"
]

[tool.pyright]
include = ["app"]
exclude = [
    "**/__pycache__",
]
defineConstant = { DEBUG = true }
reportMissingImports = "error"
reportMissingTypeStubs = false
venvPath = ".venv"
